Domain event extends SerializedEvent
